+++ credentials/java/com/google/auth/Credentials.java
@@ -71,6 +71,18 @@ public abstract class Credentials implements Serializable {
   }
 
   /**
+   * Gets the credential type used for internal metrics header.
+   *
+   * <p>The default is {@code CredentialTypeForMetrics.DO_NOT_SEND}. For a credential that is
+   * established to track for metrics, this default should be overridden.
+   *
+   * @return a enum value for credential type
+   */
+  public CredentialTypeForMetrics getMetricsCredentialType() {
+    return CredentialTypeForMetrics.DO_NOT_SEND;
+  }
+
+  /**
    * Get the current request metadata, refreshing tokens if required.
    *
    * <p>This should be called by the transport layer on each request, and the data should be